--- a/felix/labelindex/labelnamevalueindex/label_name_value_index.go
+++ b/felix/labelindex/labelnamevalueindex/label_name_value_index.go
@@ -55,7 +55,7 @@ func New[ItemID comparable, Item Labeled](nameOfTrackedItems string) *LabelNameV
 }
 
 type values[ItemID comparable] struct {
-	m     map[uniquestr.Handle]set.Set[ItemID]
+	m     map[uniquestr.Handle]*set.Adaptive[ItemID]
 	count int
 }
 
@@ -77,13 +77,13 @@ func (idx *LabelNameValueIndex[ItemID, Item]) Add(id ItemID, item Item) {
 		vals, ok := idx.labelNameToValueToIDs[k]
 		if !ok {
 			vals = values[ItemID]{
-				m: map[uniquestr.Handle]set.Set[ItemID]{},
+				m: map[uniquestr.Handle]*set.Adaptive[ItemID]{},
 			}
 			idx.labelNameToValueToIDs[k] = vals
 		}
 		setOfIDs := vals.m[v]
 		if setOfIDs == nil {
-			setOfIDs = set.New[ItemID]()
+			setOfIDs = set.NewAdaptive[ItemID]()
 			vals.m[v] = setOfIDs
 		}
 		setOfIDs.Add(id)