--- a/felix/multidict/multidict.go
+++ b/felix/multidict/multidict.go
@@ -18,7 +18,7 @@ import (
 	"github.com/projectcalico/calico/libcalico-go/lib/set"
 )
 
-type Multidict[K comparable, V comparable] map[K]set.Set[V]
+type Multidict[K comparable, V comparable] map[K]set.Typed[V]
 
 func New[K comparable, V comparable]() Multidict[K, V] {
 	return make(Multidict[K, V])