## Description

<!-- A few sentences describing the overall goals of the pull request's commits.
Please include
- the type of fix - (e.g. bug fix, new feature, documentation)
- some details on _why_ this PR should be merged
- the details of the testing you've done on it (both manual and automated)
- which components are affected by this PR
- links to issues that this PR addresses
-->

Add an "adaptive" implementation of `set.Set` that uses a different storage approach depending on number of items.

* 0 items: store a nil pointer.
* 1 item: store pointer to the item
* 2-16 items: store pointer to array of that length rounded up to power of 2
* 17+: store pointer to a proper map.

It is faster and uses a fraction of the RAM for 0-16 items, then its performance becomes similar to the normal set.Typed (with slight overhead).

Use the map in the named port index, where we have many sets that are generally expected to be very small.

It knocks 20% off the `BenchmarkSnapshot200Local250kTotal10000TagPols` benchmark, which has a very large number of small sets.  It gives almost no improvement on the others, but those are dominated by having large numbers of active policies and very few labels.

```
Before:

goos: linux
goarch: amd64
pkg: github.com/projectcalico/calico/felix/calc
cpu: Intel(R) Core(TM) i9-8950HK CPU @ 2.90GHz
BenchmarkSnapshot200Local250kTotal10000TagPols-12                     	       1	3245931822 ns/op	       307.9 HeapAllocMB	        29.00 IPSets	     10432 Msgs	     10000 Policies	         3.246 s	970140472 B/op	10684941 allocs/op
BenchmarkSnapshot200Local10kTotal1000NetsetPols-12                    	       2	 516745681 ns/op	        78.60 HeapAllocMB	      1000 IPSets	      2403 Msgs	      1000 Policies	         0.5117 s	245642908 B/op	 1909899 allocs/op
BenchmarkSnapshot200Local10kTotal10000NetsetPols-12                   	       1	5866074645 ns/op	       604.9 HeapAllocMB	     10000 IPSets	     20403 Msgs	     10000 Policies	         5.866 s	2525107568 B/op	17306981 allocs/op

After:

goos: linux
goarch: amd64
pkg: github.com/projectcalico/calico/felix/calc
cpu: Intel(R) Core(TM) i9-8950HK CPU @ 2.90GHz
BenchmarkSnapshot200Local250kTotal10000TagPols-12                     	       1	3230500160 ns/op	       249.8 HeapAllocMB	        29.00 IPSets	     10432 Msgs	     10000 Policies	         3.230 s	929163456 B/op	11895089 allocs/op
BenchmarkSnapshot200Local10kTotal1000NetsetPols-12                    	       2	 510494161 ns/op	        77.94 HeapAllocMB	      1000 IPSets	      2403 Msgs	      1000 Policies	         0.5055 s	245306080 B/op	 1962315 allocs/op
BenchmarkSnapshot200Local10kTotal10000NetsetPols-12                   	       1	5766052793 ns/op	       598.9 HeapAllocMB	     10000 IPSets	     20403 Msgs	     10000 Policies	         5.766 s	2514418280 B/op	17287133 allocs/op
```

Also included: store sets as their concrete type in the places where we store a lot of them; saves a few MB.

## Related issues/PRs

<!-- If appropriate, include a link to the issue this fixes.
fixes <ISSUE LINK>

If appropriate, add links to any number of PRs documented by this PR
documents <PR LINK>
-->
CORE-11400

## Todos

- [x] Tests
- [ ] Documentation
- [x] Release note

## Release Note

<!-- Writing a release note:
- By default, no release note action is required.
- If you're unsure whether or not your PR needs a note, ask your reviewer for guidance.
- If this PR requires a release note, update the block below to include a concise note describing
  the change and any important impacts this PR may have.
-->

```release-note
Felix now stores its selector index more efficiently resulting in reduced memory usage in large clusters.
```

## Reminder for the reviewer

Make sure that this PR has the correct labels and milestone set.

Every PR needs one `docs-*` label.

- `docs-pr-required`: This change requires a change to the documentation that has not been completed yet.
- `docs-completed`: This change has all necessary documentation completed.
- `docs-not-required`: This change has no user-facing impact and requires no docs.

Every PR needs one `release-note-*` label.

- `release-note-required`: This PR has user-facing changes. Most PRs should have this label.
- `release-note-not-required`: This PR has no user-facing changes.

Other optional labels:

- `cherry-pick-candidate`: This PR should be cherry-picked to an earlier release. For bug fixes only.
- `needs-operator-pr`: This PR is related to install and requires a corresponding change to the operator.